add support for cpu binding with srun
Greasy reads how many CPUs avail in reservation
then automatically creates strides and binds tasks
to CPUs if variable usecpubinding is set to 1. Cannot
be used if partial node is reserved, also changed
default num workers to cpus in greasy reservation/2.
Details of this process added to logs, (level 3 and 4).